+++ client/src/modules/members/components/MemberDetails.scss
